feat: restructure lot list table components and update imports
