+++ libs/client-feature-lots/src/components/lot-list-table/lot-list-table-row.tsx
