+++ libs/client-feature-lots/src/views/lots-list-view.tsx
@@ -1,6 +1,6 @@
 import LotListSummaryPanel from '../components/lot-list-summary-panel';
 import { StyledHeaderBodyContainer } from '@adg/client-theme';
-import { LotListTable } from '../components/lot-list-table.tsx/lot-list-table';
+import { LotListTable } from '../components/lot-list-table/lot-list-table';
 
 export const LotsListView: React.FC = () => {
   return (
